Example data for MsQuality: data set of Lee et al. (2019)

The data set of Lee et al. (2019) contains metabolite information measured by reverse phase liquid chromatography (RPLC) coupled to mass spectrometry and hydrophilic interaction liquid chromatography (HILIC) coupled to mass spectrometry (file `STables - rev1.xlsx` in the Supplementary Information).

It will be used as an example data set in the vignette to show the functionality of the packages. The file contains Spectra and MsExperiment objects that store the mass spectrometry data.

Format

Spectra and MsExperiment

Source

See the file Lee2019-data-source.R in scripts for the source code how sps_hilic and sps_rplc were created.

Value

Spectra and MsExperiment objects

References

Lee et al. (2019). A large-scale analysis of targeted metabolomics data from heterogeneous biological samples provides insights into metabolite dynamics. Metabolomics, 103, doi: 10.1007/s11306-019-1564-8.

Author

Thomas Naake, thomasnaake@googlemail.com