Transform the metrics into a list of MzQCmzQC objects

The function transformIntoMzQC transfers the metrics stored in spectra_metrics into a list of MzQCmzQC objects. Each list entry will refer to the corresponding entry in spectra_metrics. As such, each entry contains information from a single dataOrigin of a Spectra or Chromatograms object.

The function transformIntoMzQC is a helper function within calculateMetricsFromSpectra and calculateMetricsFromChromatograms.

transformIntoMzQC(spectra_metrics)

Arguments

spectra_metrics

list of named vector

Value

list containing as entries MzQCmzQC objects for each Spectra with same dataOrigin

Details

The MzQCmzQC object will only contain those quality metrics that have a corresponding attribute with a [PSI:MS] identifier. The matching is done via the names of each vector in spectra_metrics.

The Field "version" is set to the current version of the rmzqc package.

The entry of "MzQCanalysisSoftware" is filled with the [PSI:MS] id of MsQuality ("MS:") and the version is taken from packageDescription("MsQuality")[["Version"]].

Author

Thomas Naake, Johannes Rainer

Examples

library(MsDataHub)
library(Spectra)
library(MsQuality)
register(SerialParam())

## define file names containing spectra data for the samples
sciex_file <- c(MsDataHub::X20171016_POOL_POS_1_105.134.mzML(),
   MsDataHub::X20171016_POOL_POS_3_105.134.mzML())
#> see ?MsDataHub and browseVignettes('MsDataHub') for documentation
#> loading from cache
#> see ?MsDataHub and browseVignettes('MsDataHub') for documentation
#> loading from cache

## import the data and assign it to the spectra object
spectra <- Spectra(sciex_file)

## define the quality metrics to be calculated
metrics <- c("areaUnderTic", "chromatographyDuration", "msSignal10xChange")

## obtain the spectra_metrics object
f <- dataOrigin(spectra)
f_unique <- unique(f)
spectra_metrics <- bplapply(f_unique, function(f_unique_i) {
     calculateMetricsFromOneSampleSpectra(
         spectra = spectra[f == f_unique_i], metrics = metrics)
     }, BPPARAM = bpparam())
names(spectra_metrics) <- f_unique

## transform to mzQC format
transformIntoMzQC(spectra_metrics)
#> Warning: File '/github/home/.cache/R/ExperimentHub/4f107b8a3066_7859' has an unknown suffix. Falling back to 'MS:1000560 ! mass spectrometer file format'.
#> Warning: File '/github/home/.cache/R/ExperimentHub/4f105de1e379_7860' has an unknown suffix. Falling back to 'MS:1000560 ! mass spectrometer file format'.
#> [[1]]
#> <MzQCmzQC>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     contactAddress: NA
#>     contactName: root
#>     controlledVocabularies: list
#>     creationDate: MzQCDateTime, R6
#>     description: A mzQC document on the sample 4f107b8a3066_7859
#>     fromData: function (data, context = "MzQCmzQC") 
#>     initialize: function (version = NA_character_, creationDate = MzQCDateTime$new(), 
#>     isValid: function (context = "MzQCmzQC") 
#>     runQualities: list
#>     self: MzQCmzQC, R6
#>     setQualities: list
#>     toJSON: function (...) 
#>     version: 0.7.0
#> 
#> [[2]]
#> <MzQCmzQC>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     contactAddress: NA
#>     contactName: root
#>     controlledVocabularies: list
#>     creationDate: MzQCDateTime, R6
#>     description: A mzQC document on the sample 4f105de1e379_7860
#>     fromData: function (data, context = "MzQCmzQC") 
#>     initialize: function (version = NA_character_, creationDate = MzQCDateTime$new(), 
#>     isValid: function (context = "MzQCmzQC") 
#>     runQualities: list
#>     self: MzQCmzQC, R6
#>     setQualities: list
#>     toJSON: function (...) 
#>     version: 0.7.0
#>